+++ noodles-bam/src/bai/io.rs
@@ -1,0 +1,5 @@
+//! BAI I/O.
+
+mod reader;
+
+pub use self::reader::Reader;