+++ noodles-bam/examples/bam_index.rs
@@ -71,7 +71,7 @@ fn main() -> io::Result<()> {
     let index = builder.build(header.reference_sequences().len());
 
     let stdout = io::stdout().lock();
-    let mut writer = bai::Writer::new(stdout);
+    let mut writer = bai::io::Writer::new(stdout);
     writer.write_index(&index)?;
 
     Ok(())